[REFACTOR/#260] ReplyStatus -> Domain Layer
