+++ app/src/main/java/com/sopt/clody/presentation/utils/navigation/Route.kt
@@ -1,5 +1,6 @@
 package com.sopt.clody.presentation.utils.navigation
 
+import com.sopt.clody.domain.model.ReplyStatus
 import kotlinx.serialization.Serializable
 
 /**
